backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
PyMySQL==1.1.0
bcrypt==4.0.1
PyJWT==2.8.0
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
